refactor: remove legacy segment listing; use registry (with maxSegmentId union for safety). Tests: remove sleeps/awaits; deterministic sealing via SegmentBuildService; add prefetchCodebooksSync for synchronous codebook warmup; fix heavy recall tests. Config: expose prefetchCodebooksSync. Docs unchanged.
